Split x86-64 instruction encoding into src/x64.c
x64-codegen.c held both the instruction encoding and the IR walk, unlike
arm-codegen.c and riscv-codegen.c, which include a separate encoder file
beside them. A stale draft of that file existed but nothing compiled it,
and it defined the ModRM mode constants with values contradicting the
ones actually in use.

src/x64.c now holds the encoding: REX and ModRM construction, the byte
and dword primitives, addressing modes, patching, and the composed
forms. It cannot have the shape of arm.c, because an x86-64 instruction
has no fixed width and so there is no word to return; the helpers append
to the code section and the caller composes them. Nothing in it reaches
into the IR or the register allocator.
